+++ rust/examples/videoroom.rs
@@ -659,7 +659,10 @@ mod participant {
                     // synchronous
                     actix::spawn(async move {
                         match transport
-                            .connect(WebRtcTransportRemoteParameters { dtls_parameters })
+                            .connect(WebRtcTransportRemoteParameters {
+                                ice_parameters: None,
+                                dtls_parameters,
+                            })
                             .await
                         {
                             Ok(_) => {
@@ -718,7 +721,10 @@ mod participant {
                     // The same as producer transport, but for consumer transport
                     actix::spawn(async move {
                         match transport
-                            .connect(WebRtcTransportRemoteParameters { dtls_parameters })
+                            .connect(WebRtcTransportRemoteParameters {
+                                ice_parameters: None,
+                                dtls_parameters,
+                            })
                             .await
                         {
                             Ok(_) => {