+++ rust/src/router/webrtc_transport.rs
@@ -129,6 +129,10 @@ pub struct WebRtcTransportOptions {
     /// Prefer TCP.
     /// Default false.
     pub prefer_tcp: bool,
+    /// ICE consent timeout (in seconds). If 0 it is disabled.
+    /// For it to be enabled, iceParameters must be given in connect() method.
+    /// Default 30.
+    pub ice_consent_timeout: u8,
     /// Create a SCTP association.
     /// Default false.
     pub enable_sctp: bool,
@@ -155,6 +159,7 @@ impl WebRtcTransportOptions {
             enable_tcp: false,
             prefer_udp: false,
             prefer_tcp: false,
+            ice_consent_timeout: 30,
             enable_sctp: false,
             num_sctp_streams: NumSctpStreams::default(),
             max_sctp_message_size: 262_144,
@@ -172,6 +177,7 @@ impl WebRtcTransportOptions {
             enable_tcp: true,
             prefer_udp: false,
             prefer_tcp: false,
+            ice_consent_timeout: 30,
             enable_sctp: false,
             num_sctp_streams: NumSctpStreams::default(),
             max_sctp_message_size: 262_144,
@@ -384,6 +390,8 @@ impl WebRtcTransportStat {
 #[derive(Debug, Clone, PartialOrd, Eq, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct WebRtcTransportRemoteParameters {
+    /// Remote ICE parameters.
+    pub ice_parameters: Option<IceParameters>,
     /// Remote DTLS parameters.
     pub dtls_parameters: DtlsParameters,
 }
@@ -920,7 +928,7 @@ impl WebRtcTransport {
     ///
     /// # Example
     /// ```rust
-    /// use mediasoup::data_structures::{DtlsParameters, DtlsRole, DtlsFingerprint};
+    /// use mediasoup::data_structures::{IceParameters, DtlsParameters, DtlsRole, DtlsFingerprint};
     /// use mediasoup::webrtc_transport::WebRtcTransportRemoteParameters;
     ///
     /// # async fn f(
@@ -929,6 +937,11 @@ impl WebRtcTransport {
     /// // Calling connect() on a PlainTransport created with comedia and rtcp_mux set.
     /// webrtc_transport
     ///     .connect(WebRtcTransportRemoteParameters {
+    ///         ice_parameters: Some(IceParameters {
+    ///             username_fragment: "foo".to_string(),
+    ///             password: "xxxx".to_string(),
+    ///             ice_lite: None,
+    ///         }),
     ///         dtls_parameters: DtlsParameters {
     ///             role: DtlsRole::Server,
     ///             fingerprints: vec![
@@ -958,6 +971,7 @@ impl WebRtcTransport {
             .request(
                 self.id(),
                 WebRtcTransportConnectRequest {
+                    ice_parameters: remote_parameters.ice_parameters,
                     dtls_parameters: remote_parameters.dtls_parameters,
                 },
             )